backend/requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.32.1
SQLAlchemy==2.0.36
pydantic==2.9.2
pydantic-settings==2.6.1
python-multipart==0.0.17
Pillow==11.0.0
